TEST: Migrate AbstractTicketProvider tests from Feature to Unit directory
